Fixed Bug for plotting
There was a bug that if a commodity is not consumed by a storage or a process over the entire span of the calculation, the "consumed" array would be empty and the stackplot function would raise an error. Hereby, it is important to note that "demand" is not part of "consumed" but plotted separately.
Added a default zero-column if "consumed" is empty